refactoring home page
