+++ Client/src/pages/home/SliderBanner.jsx
@@ -1,7 +1,7 @@
 import 'slick-carousel/slick/slick.css';
 import 'slick-carousel/slick/slick-theme.css';
 import Slider from 'react-slick';
-import './SliderBanner.css';
+import './components/SliderBanner.css';
 import {useState} from 'react';
 
 const SliderBanner = () => {
